requirements.txt
pycryptodome 
pysocks
stem
websockets
Flask
requests
